Custom header with account settings drop-down
Updated the custom header to include an app icon and the user's name and initials (currently hard-coded with my name). The user name is clickable and opens a drop-down menu with links to account settings and to log out.

In other news, I added a favicon for the app and included the jQuery library.