Calendar create event button
Needed to reorder the calendar's navigation UI a bit to add the create button. Also improved the integration with the database (now loads all of a user's course sessions). Finally, added to the Database class so that it automatically attempts to reconnect to MySQL if the connection is lost (in the isConnected() method).