Okay adding the scheduler class and update to some of priority class
So I would call iteration 1 of the scheduler class mostly done. It works with weekly scheduled events and I'm going to test it with other study sessions scheduled tomorrow as well as once off events. But the basis is there and it is working. the rest must just be added as checks before writing the study session. I don't know how but somehow the design of this class became very modular. Long description but I worked hard on this and I deserve it so nahhhh