Semesters page finished!
New users (i.e. users with no courses in the database yet) are automatically redirected to the Semesters page upon opening the app. If they try to create a session from the Calendar page, they are prompted to first create some semesters and courses on the Semesters page. There is one small bug in Jess's delete method, but I will solve that tonight. Will try to integrate Tasks before bed.